refactor(diffusion): move ring attention primitives to shared modules
_ring_attention_varlen (P2P-rotated varlen ring with online-softmax merge)
and _ring_merge_attention move from the MiniMax H3 model into layers/usp.py
next to ring_attn, decoupled from the model via an explicit softmax_scale;
the (world_size, rank) convenience getters with uninitialized fallbacks move
into distributed/parallel_state.py as get_ulysses_ctx/get_ring_ctx.
